refactor(ppp): split DdFilterScaffold out of ppp_clas_dd.cpp
- ppp_clas_dd_internal.hpp: former anonymous-namespace helpers (DD group
  keys, candidate structs, layout math) as inline functions in
  libgnss::ppp_clas_dd::internal
- ppp_clas_dd_scaffold.cpp: the DdFilterScaffold snapshot/predict/hold
  implementation (~1.5k lines)
- ppp_clas_dd.cpp keeps StateLayout, DD measurement building and postfit
  validation

Verified: ctest 71/71, byte-identical rtk_positioning output on PPC
tokyo/run1, MADOCALIB parity suite 17/17.
